requirements.txt
# Core RAG dependencies
langchain==0.3.7
langchain-community==0.3.6
faiss-cpu==1.8.0

# Hugging Face for FREE cloud deployment
huggingface-hub==0.20.3

# Ollama for local LLMs (Llama 3, Mistral, etc.)
ollama==0.4.5

# Local embeddings (free, no API needed)
sentence-transformers==3.3.1

# Document processing
pypdf==4.3.1
python-docx==1.1.2
pandas==2.2.3
openpyxl==3.1.5

# API & UI
fastapi==0.115.4
uvicorn[standard]==0.32.0
gradio==4.44.1
pydantic==2.9.2
pydantic-settings==2.6.0

# Utilities
python-dotenv==1.0.1
tqdm==4.66.5
numpy==1.26.4
requests==2.31.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
